refactor: Rename token to sync
